backend-service/requirements.txt
Flask==3.0.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
gunicorn==21.2.0
python-dotenv==1.0.0

